Uploaded my edited file for the alphabetical testing
Uploaded my edited file for the alphabetical testing  xlsm. It now includes my final version of the VBA code included in the VBA module, i believe it is under sub fillcolumns(). But this has my final version of the macro saved within this file if you would need to check it out and the readme isn't executable